src/backend/requirements.txt
asgiref==3.4.1
certifi==2021.5.30
click==8.0.1
fastapi==0.68.1
h11==0.12.0
minio==7.1.0
pydantic==1.8.2
python-multipart==0.0.5
six==1.16.0
starlette==0.14.2
typing-extensions==3.10.0.2
urllib3==1.26.7
uvicorn==0.15.0
python-dotenv
psycopg2-binary
pandas
python-chess